backend_service/requirements.txt
fastapi
uvicorn[standard]
pymongo
dnspython
certifi
joblib
xgboost
scikit-learn
pandas
numpy
python-dotenv
scipy
